requirements.txt
line-bot-sdk
future
Flask
click
Werkzeug
jsonschema
unicodecsv
urllib3
requests
